Move some interfaces out of types.go make it simpler
